+++ lib.php
@@ -219,6 +219,7 @@ function confsubmissions_delete_instance($id) {
     $DB->delete_records('confsubmissions_submission', ['confsubmissions' => $id]);
     $DB->delete_records('confsubmissions_field', ['confsubmissions' => $id]);
     $DB->delete_records('confsubmissions_track', ['confsubmissions' => $id]);
+    $DB->delete_records('confsubmissions_submissiontype', ['confsubmissions' => $id]);
 
     $DB->delete_records('confsubmissions', ['id' => $id]);